Revert "feat: 상세 조회"
